move some to array
